rotando pero parece que se mueve

+++ includes/cub3D.h
@@ -116,6 +116,6 @@ void	exit_cub(t_args *args);
 void	print_vert_line(int x, int ystart, int yend, int color, t_args *args);
 void	print_walls(t_args *args, t_map *map, t_player *pl);
 void	init_player(t_map	*map);
-void	rotate(t_player *pl, float a);
+void	rotate(t_player *pl, double a);
 
 #endif
